refactor - shifted internal native api interfaces into internals - corrected compile errors that resulted
To Do:
- unit tests still need updated to build
