+++ CSharp/CredentialStore/NativeApi/ICriticalCredentialHandle.cs
@@ -14,7 +14,7 @@
 
 using System;
 
-namespace Moreland.Security.Win32.CredentialStore
+namespace Moreland.Security.Win32.CredentialStore.NativeApi
 {
     public interface ICriticalCredentialHandle : IDisposable
     {
