+++ CSharp/CredentialStore/NativeApi/IMarshalService.cs
@@ -12,15 +12,17 @@
 // 
 
 using System;
+#if !NETSTANDARD2_0
 using System.Diagnostics.CodeAnalysis;
+#endif
 using System.Runtime.InteropServices;
 
-namespace Moreland.Security.Win32.CredentialStore
+namespace Moreland.Security.Win32.CredentialStore.NativeApi
 {
     /// <summary>
     /// Wrapper around <see cref="Marshal"/>
     /// </summary>
-    public interface IMarshalService
+    internal interface IMarshalService
     {
         /// <summary>
         /// <inheritdoc cref="Marshal.PtrToStructure(IntPtr, Type)"/>
